requirements.txt
google-genai
playwright
python-dotenv
